backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
langchain>=0.2.0
langchain-google-genai>=1.0.5
langchain-community>=0.2.0
chromadb>=0.5.0
python-docx>=1.1.2
python-pptx>=0.6.23
reportlab>=4.2.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
